Copy 'default' to 'example' in specs and format the files
'Default' should really not be used in these cases according to the spec
as it implies the value used when a parameter is NOT provided. We should
really use 'example'. The problem is that the Readme platform requires
the use of 'default' for the generated code examples to use these values
for the 'Try it out' button. So we keep both.